requirements.txt
# Pur Ferme Backend
flask>=3.0.0
flask-cors>=4.0.0
torch>=2.1.0
torchvision>=0.16.0
fastai>=2.7.14
Pillow>=10.0.0
requests>=2.31.0
python-dotenv>=1.0.0
google-generativeai>=0.7.0
openai>=1.30.0